+++ src/main/java/com/example/goodfood/entity/Category.java
@@ -14,7 +14,7 @@ import javax.persistence.*;
 public class Category {
     @Id @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "cate_id", nullable = false)
-    private int cateId;
+    private Long cateId;
     @Column(name = "cate_name", nullable = true)
     private String cateName;
